features/steps/device_lib/requirements.txt
bitstring==3.1.9
boto3==1.24.70
botocore==1.27.70
cffi==1.15.1
cryptography==38.0.1
ecdsa==0.18.0
esptool==4.2.1
jmespath==1.0.1
pycparser==2.21
pyserial==3.5
pynput==1.7.6
python-dateutil==2.8.2
reedsolo==1.5.4
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
pyinstaller==5.8.0